Rename useBrassring filename to match hook name
